Add files via upload
There is a M-M relationship between staff and certifications–the associative entity created is certifications_obtained. A staff can have many certifications and a certificate can be obtained by many individuals. There is a 1-M non-identifying relationship between staff and class_type. A staff can teach multiple classes and a class has one default instructor(staff). There is a 1-M relationship between staff and class_session; a staff can be an instructor for many classes. Another 1-M relationship between staff and class_session shows that a staff can also be a fitness monitor for many classes. Lastly, there is a 1-M identifying relationship between staff and pt_session showing that a staff member can be an instructor for many personal training sessions. There is also a 1-M identifying relationship between pt_session and client. A personal training session can only be created if an instructor and client schedule the session.
Class_session has a 1-1 identifying relationship with class_type; a sports session is identified based on the type of class it is. There is also a M-M identifying relationship between class_session and client–the associative entity between the two is class_attendance. The records of class attendance is dependent on the type of class a session is and the number of clients that attend each session.